♻️ 4Z7EZP loops: split executable runtime
